Add shared functions for feed_dict filling
